Update map loader tests and move old code into tools
